python/requirements.txt
Adafruit-Blinka==8.28.1
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-neopixel==6.3.11
adafruit-circuitpython-pixelbuf==2.0.4
adafruit-circuitpython-requests==2.0.3
adafruit-circuitpython-typing==1.9.6
Adafruit-PlatformDetect==3.57.0
Adafruit-PureIO==1.1.11
board==1.0
contourpy==1.2.0
cycler==0.12.1
fonttools==4.47.0
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.2
Pillow==8.1.2
pyftdi==0.55.0
pyparsing==3.1.1
pyserial==3.5
python-dateutil==2.8.2
PyTurboJPEG @ git+https://github.com/lilohuang/PyTurboJPEG.git@c9a4973ab48e1e3d421881f76d8b1b2a22669fd2
pyusb==1.2.1
rpi-ws281x==5.0.0
scipy==1.11.4
six==1.16.0
typing-extensions==4.9.0
v4l2-python3==0.3.2
v4l2py==2.3.0
